pipeline: allow re-setting video caps
The very first FIFO target added to the pipeline decides what the output
video resolution will be, any subsequent FIFOs only inherit that value
and their own resolution setting is ignored.

When all FIFOs are removed (self->target becomes empty), give the FIFO
that gets added next a chance to re-set the stream resolution to some
new value.

This also shows the API of GaeguliPipeline isn't well defined. Either
the resolution should be a property of the whole GaeguliPipeline, not
individual targets, or each target pipeline should include a videoscale
element that will produce stream in that target's requested resolution.