pipeline: make it independent of encoding method
The encoding method doesn't need to be known on the level of the whole
pipeline. Remove that parameter from GaeguliPipeline constructor and
require it to be specified when adding a new target.

The only decision GaeguliPipeline did based on the selected encoding
method was whether decodebin and clockoverlay should be added into the
vsrc pipeline. This change uses decodebin at all times unless the input
device is nvarguscamerasrc, which produces video/x-raw(memory:NVMM)
video frames.