target: fix pipeline hang when removing a target
Setting the state of srtsink to be disposed to NULL will free the
occupied UDP port immediately, but passing a buffer to a sink in
GST_STATE_NULL will return GST_FLOW_FLUSHING, which could in turn pause
our video source and disturb other streaming targets.

To prevent this, install a pad probe before the srtsink that drops all
buffers.